requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
filetype==1.2.0
firecrawl-py==3.3.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langsmith==0.4.16
MarkupSafe==3.0.2
multidict==6.6.4
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.3.2
orjson==3.11.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.0
rsa==4.9.1
serpapi==0.1.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
streamlit==1.48.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.24.0